update and delete ownership

+++ controllers/posts.js
@@ -27,18 +27,34 @@ export const getSinglePost = asyncHandler(async (req, res) => {
 
 export const updatePost = asyncHandler(async (req, res) => {
   const {
+    user,
     params: { id },
     body: { title, body, author }
   } = req;
-  const updatedPost = await Post.findOneAndUpdate({ _id: id }, { title, body, author }, { new: true });
-  if (!updatedPost) throw new ErrorResponse(`Post with id of ${id} not found, could not update`, 404);
+  const postToUpdate = await Post.findById(id);
+  if (!postToUpdate)
+    throw new ErrorResponse(`Post with id of ${id} not found, could not update`, 404);
+  if (user.id !== postToUpdate.author.toString())
+    throw new ErrorResponse(`Only the author of the post can modify it. Go away`, 403);
+
+  const updatedPost = await Post.findOneAndUpdate(
+    { _id: id },
+    { title, body, author },
+    { new: true }
+  );
   res.json(updatedPost);
 });
 
 export const deletePost = asyncHandler(async (req, res) => {
   const {
+    user,
     params: { id }
   } = req;
+  const postToUpdate = await Post.findById(id);
+  if (!postToUpdate)
+    throw new ErrorResponse(`Post with id of ${id} not found, could not update`, 404);
+  if (user.id !== postToUpdate.author.toString())
+    throw new ErrorResponse(`Only the author of the post can modify it. Go away`, 403);
   const deleted = await Post.findOneAndDelete({ _id: id });
   if (!deleted) throw new ErrorResponse(`Post with id of ${id} not found, could not delete`, 404);
   res.json({ success: `Post with id of ${id} was deleted` });
